Add support for retrieving unique locations and make values
This commit adds support for retrieving unique locations and make values in the API. The code now includes conditional checks for the "unique" query parameter and handles different values of "unique" accordingly. The response is formatted as JSON and appropriate HTTP response codes are set based on the success or failure of the operations.