Add ComputersTable component
A new component named 'ComputersTable' is created. It is a React function component designed to fetch and display a list of computers from a remote API. Furthermore, a 'Computer' type definition is added for type safety, and an 'Eye' function that adjusts its color based on the current theme is also included.