Refactor Home component to use ComputersTable
The Home component has been refactored to use the new ComputersTable component, removing a large amount of redundant code and improving maintainability. The eye function, which was unused, has also been removed to clean up the code. This update enhances the structure of the codebase making it easier to manage.